Honor colorspaces "DeviceGray" and "DeviceRGB"
Improved checking if the raw pixel data fits to the actual file contents
and memorization of color spaces. Apparently it is possible to have an
DeviceRGB color space image that is only 1/3rd of the expected size
(i.e., de-facto grayscale, but probably interpreted as RGB?). Handling
this as a conditional depending on the pixel data size right now, but
might not be the correct solution.