Implement T1 font decoding
T1 font encoding is required to get access to the CharSet array that
needs to be set in the T1 font descriptor of a PDF/A-1b compliant
document. This sets the prerequisite for this. Parsing of the T1 font
isn't done using a PostScript interpreter, however, and therefore
possibly cannot decode all fonts correctly.