Generic stream decompression and deprediction
Prediction is a PDF filter that is used before objects are compressed.
Previously, only direct data was handled. With the new EncodeDecode
facility, generic decompression and/or deprediction is possible now.
Ultimately, this should replace the .stream property of PDFFile and what
is now known as .stream should become .raw_stream.