Signature now works in both Acrobat and X-Viewer
Signatures now are validatable in both X-Viewer and Adobe Acrobat.
However, I think that in X-Viewer the sig does not globally show up (but
works when clicked on the annotation). Therefore I'll need to fix this
too. Note that the signature does not work anymore in Acrobat when
removing the Locked flag from the /F key.