Implement reading PFB files
Read PFB files and convert them into T1 /FontFile objects. This is
needed because if we want to sign an object, we want to display some
font in the form that is for verification. However, when also trying to
create a PDF/A-1b, fonts need to be embedded. Therefore we need to
include our own font. PFB is easiest to handle.