Test run of signing which manually inserts font
For now, this works on my machine (where I have manually extracted the
PDF Font Descriptor from a pdfLaTeX document). However, the ultimate
solution would be to let the user supply a PBF and create the FontFile
and FontDescriptor from that. Since this is not done yet, signing is
still somewhat broken for now.