Character mappings for seac
This somehow really doesn't work out as nicely as it should. seac should
indicate the name of the base and name of the accented character.
However, for example for Adobe Courier we get /Aacute with values of 65
("/A") and 194 ("/Acircumflex"). My expectation would have been to see
193 ("/Aacute"), but that's not the case (t1disas confirms that the
decoding is done correctly). So I must have a misunderstanding of how
seac internally works. I'll figure this out on another day.