Refactoring of PDFResources
PDFResources is now known as PDFTemplate and has a clean exterior
interface. This means that all PDFTemplates need to be self-contained
(i.e., provide no non-exported dangling objects and have no unresolved
input references -- both are checked during coalescing with the
PDFFile). Also include the beautiful Type1 font Bitstream Charter so
that a font is available if the user is unwilling/unable to provide a
PFB file by themselves.