Enable token parser to read until EOF
In order to fix #4, we must be able to read tokens until EOF, since the
initial problem with the PDF provided by @altblue is that there is no
token delimiter (e.g., newline) after the final "%%EOF" token. However,
since EOF isn't a bytes() object, we had to rework the parser quite a
bit to get the job done. Note that #4 isn't fully fixed with this, since
there's a separate issue with the provided PDF that I'll need to look
into next.