Fix indirect access of Page/Resources object
For the example PDF provided by @altblue, the /Resources object was
directly encoded instead as a XReference. Ideally, we should find a way
to generially dereference such objects (i.e., automatically following
all XRef links). For now, we simply recognize this special case (but
that same problem may lurk in other places as well, so it's a shoddy
fix). Also include the test PDF in the test suite for the future. This
commit finally fixes #4.